+++ pkg/reconciler/v1alpha1/namespace/namespace.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2018 The Knative Authors
+Copyright 2019 The Knative Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -18,25 +18,113 @@ package namespace
 
 import (
 	"context"
-
 	"github.com/knative/eventing/contrib/gcppubsub/pkg/util/logging"
 	"github.com/knative/eventing/pkg/apis/eventing/v1alpha1"
 	"go.uber.org/zap"
+	"k8s.io/api/core/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/dynamic"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/record"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/controller"
+	"sigs.k8s.io/controller-runtime/pkg/handler"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+	"sigs.k8s.io/controller-runtime/pkg/source"
 )
 
 const (
+	// controllerAgentName is the string used by this controller to identify
+	// itself when creating events.
+	controllerAgentName = "knative-eventing-namespace-controller"
+
 	defaultBroker             = "default"
 	knativeEventingAnnotation = "eventing.knative.dev/injection"
 
 	// Name of the corev1.Events emitted from the reconciliation process
 	brokerCreated = "BrokerCreated"
 )
+
+type reconciler struct {
+	client        client.Client
+	restConfig    *rest.Config
+	dynamicClient dynamic.Interface
+	recorder      record.EventRecorder
+
+	logger *zap.Logger
+}
+
+// Verify the struct implements reconcile.Reconciler
+var _ reconcile.Reconciler = &reconciler{}
+
+// ProvideController returns a function that returns a Broker controller.
+func ProvideController(logger *zap.Logger) func(manager.Manager) (controller.Controller, error) {
+	return func(mgr manager.Manager) (controller.Controller, error) {
+		// Setup a new controller to Reconcile Brokers.
+		r := &reconciler{
+			recorder: mgr.GetRecorder(controllerAgentName),
+			logger:   logger,
+		}
+		c, err := controller.New(controllerAgentName, mgr, controller.Options{
+			Reconciler: r,
+		})
+		if err != nil {
+			return nil, err
+		}
+
+		// Watch Namespaces.
+		if err = c.Watch(&source.Kind{Type: &v1.Namespace{}}, &handler.EnqueueRequestForObject{}); err != nil {
+			return nil, err
+		}
+
+		// Watch all the resources that this reconciler reconciles.
+		if err = c.Watch(&source.Kind{Type: &v1alpha1.Broker{}}, &handler.EnqueueRequestsFromMapFunc{ToRequests: &namespaceMapper{}}); err != nil {
+			return nil, err
+		}
+
+		for _, t := range []runtime.Object{ &v1alpha1.Broker{} } {
+			err = c.Watch(&source.Kind{Type: t}, &handler.EnqueueRequestsFromMapFunc{ToRequests: &namespaceMapper{}});
+			if err != nil {
+				return nil, err
+			}
+		}
+
+		return c, nil
+	}
+}
+
+type namespaceMapper struct{}
+
+var _ handler.Mapper = &namespaceMapper{}
+
+func (namespaceMapper) Map(o handler.MapObject) []reconcile.Request {
+	return []reconcile.Request{
+		{
+			NamespacedName: types.NamespacedName{
+				Namespace: "",
+				Name:      o.Meta.GetNamespace(),
+			},
+		},
+	}
+}
+
+func (r *reconciler) InjectClient(c client.Client) error {
+	r.client = c
+	return nil
+}
+
+func (r *reconciler) InjectConfig(c *rest.Config) error {
+	r.restConfig = c
+	var err error
+	r.dynamicClient, err = dynamic.NewForConfig(c)
+	return err
+}
 
 // Reconcile compares the actual state with the desired, and attempts to
 // converge the two. It then updates the Status block of the Trigger resource
@@ -103,7 +191,7 @@ func (r *reconciler) getBroker(ctx context.Context, ns *corev1.Namespace) (*v1al
 func (r *reconciler) reconcileBroker(ctx context.Context, ns *corev1.Namespace) (*v1alpha1.Broker, error) {
 	current, err := r.getBroker(ctx, ns)
 
-	// If the resource doesn't exist, we'll create it
+	// If the resource doesn't exist, we'll create it.
 	if k8serrors.IsNotFound(err) {
 		b := newBroker(ns)
 		err = r.client.Create(ctx, b)
